requirements.txt
asgiref==3.5.2
boto3==1.24.32
botocore==1.27.32
cloudinary==1.29.0
coverage==6.4.2
crispy-bootstrap5==0.6
dj-database-url==0.5.0
dj3-cloudinary-storage==0.0.6
Django==3.2.13
django-allauth==0.51.0
django-amazon-ses==4.0.1
django-cloudinary-storage==0.3.0
django-crispy-forms==1.14.0
django-js-asset==2.0.0
django-ses==3.1.0
django-summernote==0.8.20.0
django-taggit==3.0.0
django-widget-tweaks==1.4.12
gunicorn==20.1.0
jmespath==1.0.1
model-bakery==1.6.0
oauthlib==3.2.0
Pillow==9.2.0
psycopg2==2.9.3
psycopg2-binary==2.8.4
PyGreSQL==5.2.4
PyJWT==2.4.0
python3-openid==3.2.0
pytz==2022.1
requests-oauthlib==1.3.1
s3transfer==0.6.0
sqlparse==0.4.2
